+++ aws_lambda_powertools/event_handler/__init__.py
@@ -8,12 +8,12 @@ from aws_lambda_powertools.event_handler.api_gateway import (
     ApiGatewayResolver,
     APIGatewayRestResolver,
     CORSConfig,
+    Response,
 )
 from aws_lambda_powertools.event_handler.appsync import AppSyncResolver
 from aws_lambda_powertools.event_handler.lambda_function_url import (
     LambdaFunctionUrlResolver,
 )
-from aws_lambda_powertools.event_handler.response import Response
 from aws_lambda_powertools.event_handler.vpc_lattice import VPCLatticeResolver, VPCLatticeV2Resolver
 
 __all__ = [
