+++ aws_lambda_powertools/event_handler/openapi/encoders.py
@@ -29,9 +29,8 @@ def decimal_encoder(dec_value: Decimal) -> Union[int, float]:
     Encodes a Decimal as int of there's no exponent, otherwise float
 
     This is useful when we use ConstrainedDecimal to represent Numeric(x,0)
-    where a integer (but not int typed) is used. Encoding this as a float
+    where an integer (but not int typed) is used. Encoding this as a float
     results in failed round-tripping between encode and parse.
-    Our Id type is a prime example of this.
 
     >>> decimal_encoder(Decimal("1.0"))
     1.0
